requirements.txt
Pillow==10.3.0
streamlit==1.34.0
transformers==4.41.0
